% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvcR.R
\name{gvcr_data}
\alias{gvcr_data}
\title{gvcr_data}
\usage{
gvcr_data(
  locations = gvc_location,
  risks = gvc_risk,
  levels = gvc_level,
  category = gvc_category
)
}
\arguments{
\item{locations}{Location of the risk.}

\item{risks}{Global value chain risk.}

\item{levels}{Level of the risk.}

\item{category}{Category of the risk.}
}
\value{
Data for the location, risk, level and risk category requested.
}
\description{
This function allows you to find and display the Global Value Chains Risks data according to the selected parameters.
If no arguments are filled, all data will be displayed.
}
\examples{

myData<- gvcr_data(location = "Panama Canal", risks = "S-T")
myData<- gvcr_data(location = c("Panama Canal", "Suez Canal"), risks = c("S-T", "S-C"))
myData<- gvcr_data(category = "Weather and climate risk")
myData<- gvcr_data(levels = "High risk")
myData<- gvcr_data("Panama Canal", "S-T")
myData<- gvcr_data()
}
\seealso{
\code{\link{gvcr_location}} for the locations list and \code{\link{gvcr_risk}} for the code of a risk.
}
