% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvcR.R
\name{gvcr_risk}
\alias{gvcr_risk}
\title{gvcr_risk}
\usage{
gvcr_risk(risk)
}
\arguments{
\item{risk}{The global value chains risk.}
}
\value{
The code of the global value chains risk.
}
\description{
This function allows you to find and search the right indicator code from the Social Progress Index you want to use.
If no argument is filed, all indicators will be displayed.
}
\examples{
myIndicator <- gvcr_risk()
myIndicator <- gvcr_risk(risk = "storm")
myIndicator <- gvcr_risk("storm")
}
\seealso{
\code{\link{gvcr_location}} for the locations list and \code{\link{gvcr_data}} to collect data.
}
